Error when generating mapper for abstract class
When an abstract class has only abstract getter and/or setter for a property, the generation fails because it can find the field or method to use to serialize or deserialize this property.
To fix this, we just skip all properties when the class is abstract (like we do for interface).