Support for Object and Serializable
With this change, it is now possible to use Object or Serializable in your bean.
By default, only the types declared in AbstractConfiguration are generated.
If you need more types, you can add them with the new method AbstractConfiguration.whitelist(String). This method takes a regex with the same format as 'rpc.blacklist'.