Basic support for JavaScriptObject
Only support the basic stringify and JsonUtils.safeEval.
Later, it will be possible to add any jackson annotation on them. The generator will detect it and will consider the JSO as any other bean. The only difference should be the default constructor.
There is also some edge case  around JsArray. What if the JSO contains in JsArray has jackson annotations ?